更新 .gitignore，添加 output/ 目录；修改 TextImg.py 中字符背景尺寸计算方式；优化 select_by_kp.pyx 中数据类型定义；在 run.py 中启用调试模式；修正 service/__init__.py 中 YAML 导入方式；改进 font_tool.py 中字符检查逻辑

+++ libs/cxx/select_by_kp.pyx
@@ -63,7 +63,7 @@ def nms(np.ndarray[np.float32_t, ndim=2] dets, np.float thresh):
 
     cdef int ndets = dets.shape[0]
     cdef np.ndarray[np.int_t, ndim=1] suppressed = \
-        np.zeros((ndets), dtype=np.int)
+        np.zeros((ndets), dtype=int)
 
     # nominal indices
     cdef int _i, _j
@@ -105,7 +105,7 @@ def nms(np.ndarray[np.float32_t, ndim=2] dets, np.float thresh):
     return keep
 
 
-cdef inline np.int vertical_similarity(np.ndarray[np.float32_t, ndim=1] box1, np.ndarray[np.float32_t, ndim=1] box2, np.float iou_thresh):
+cdef inline int vertical_similarity(np.ndarray[np.float32_t, ndim=1] box1, np.ndarray[np.float32_t, ndim=1] box2, np.float iou_thresh):
 
     cdef float intersec_height = min(box1[3], box2[3]) - max(box1[1], box2[1])
     cdef float union_height = max(box1[3], box2[3]) - min(box1[1], box2[1])
@@ -116,7 +116,7 @@ cdef inline np.int vertical_similarity(np.ndarray[np.float32_t, ndim=1] box1, np
         return 0####
 
 
-def box_connect(np.ndarray[np.float32_t, ndim=2] imageIntegral, np.ndarray[np.float32_t, ndim=2] anchors_in, np.int h_gap, np.int width, np.int height):
+def box_connect(np.ndarray[np.float32_t, ndim=2] imageIntegral, np.ndarray[np.float32_t, ndim=2] anchors_in, int h_gap, int width, int height):
     """
 
     :param imageIntegral:
